move `RustIrSource` impl to the database
