+++ src/db.rs
@@ -1,5 +1,3 @@
-#![allow(non_camel_case_types)]
-
 use crate::error::ChalkError;
 use crate::lowering::LowerGoal;
 use crate::program::Program;
@@ -8,9 +6,23 @@ use chalk_ir::could_match::CouldMatch;
 use chalk_ir::tls;
 use chalk_ir::DomainGoal;
 use chalk_ir::Goal;
+use chalk_ir::Identifier;
+use chalk_ir::ImplId;
 use chalk_ir::IsCoinductive;
+use chalk_ir::Parameter;
 use chalk_ir::ProgramClause;
+use chalk_ir::ProjectionTy;
+use chalk_ir::StructId;
 use chalk_ir::TraitId;
+use chalk_ir::Ty;
+use chalk_ir::TypeId;
+use chalk_ir::TypeKindId;
+use chalk_ir::TypeName;
+use chalk_rules::RustIrSource;
+use chalk_rust_ir::AssociatedTyDatum;
+use chalk_rust_ir::ImplDatum;
+use chalk_rust_ir::StructDatum;
+use chalk_rust_ir::TraitDatum;
 use chalk_solve::ProgramClauseSet;
 use chalk_solve::SolverChoice;
 use salsa::Database;
@@ -71,5 +83,72 @@ impl IsCoinductive for ChalkDatabase {
         } else {
             false
         }
+    }
+}
+
+impl RustIrSource for ChalkDatabase {
+    fn associated_ty_data(&self, ty: TypeId) -> Arc<AssociatedTyDatum> {
+        self.program_ir().unwrap().associated_ty_data[&ty].clone()
+    }
+
+    fn trait_datum(&self, id: TraitId) -> Arc<TraitDatum> {
+        self.program_ir().unwrap().trait_data[&id].clone()
+    }
+
+    fn impl_datum(&self, id: ImplId) -> Arc<ImplDatum> {
+        self.program_ir().unwrap().impl_data[&id].clone()
+    }
+
+    fn struct_datum(&self, id: StructId) -> Arc<StructDatum> {
+        self.program_ir().unwrap().struct_data[&id].clone()
+    }
+
+    fn impls_for_trait(&self, trait_id: TraitId) -> Vec<ImplId> {
+        self.program_ir()
+            .unwrap()
+            .impl_data
+            .iter()
+            .filter(|(_, impl_datum)| {
+                let impl_trait_id = impl_datum.binders.value.trait_ref.trait_ref().trait_id;
+                impl_trait_id == trait_id
+            })
+            .map(|(&impl_id, _)| impl_id)
+            .collect()
+    }
+
+    fn impl_provided_for(&self, auto_trait_id: TraitId, struct_id: StructId) -> bool {
+        // Look for an impl like `impl Send for Foo` where `Foo` is
+        // the struct.  See `push_auto_trait_impls` for more.
+        let type_kind_id = TypeKindId::StructId(struct_id);
+        self.program_ir()
+            .unwrap()
+            .impl_data
+            .values()
+            .any(|impl_datum| {
+                let impl_trait_ref = impl_datum.binders.value.trait_ref.trait_ref();
+                impl_trait_ref.trait_id == auto_trait_id
+                    && match impl_trait_ref.parameters[0].assert_ty_ref() {
+                        Ty::Apply(apply) => match apply.name {
+                            TypeName::TypeKindId(id) => id == type_kind_id,
+                            _ => false,
+                        },
+
+                        _ => false,
+                    }
+            })
+    }
+
+    fn type_name(&self, id: TypeKindId) -> Identifier {
+        match self.program_ir().unwrap().type_kinds.get(&id) {
+            Some(v) => v.name,
+            None => panic!("no type with id `{:?}`", id),
+        }
+    }
+
+    fn split_projection<'p>(
+        &self,
+        projection: &'p ProjectionTy,
+    ) -> (Arc<AssociatedTyDatum>, &'p [Parameter], &'p [Parameter]) {
+        self.program_ir().unwrap().split_projection(projection)
     }
 }
