+++ src/program.rs
@@ -1,12 +1,9 @@
-use crate::program_environment::ProgramEnvironment;
 use chalk_ir::debug::Angle;
 use chalk_ir::tls;
 use chalk_ir::{
-    Identifier, ImplId, Parameter, ProgramClause, ProjectionTy, StructId, TraitId, Ty, TypeId,
-    TypeKindId, TypeName,
+    Identifier, ImplId, Parameter, ProgramClause, ProjectionTy, StructId, TraitId, TypeId,
+    TypeKindId,
 };
-use chalk_rules::clauses::ToProgramClauses;
-use chalk_rules::RustIrSource;
 use chalk_rust_ir::{AssociatedTyDatum, ImplDatum, ImplType, StructDatum, TraitDatum, TypeKind};
 use std::collections::BTreeMap;
 use std::fmt;
@@ -110,133 +107,5 @@ impl tls::DebugContext for Program {
             associated_ty_data.name,
             Angle(&other_params)
         )
-    }
-}
-
-impl RustIrSource for Program {
-    fn associated_ty_data(&self, ty: TypeId) -> Arc<AssociatedTyDatum> {
-        self.associated_ty_data[&ty].clone()
-    }
-
-    fn trait_datum(&self, id: TraitId) -> Arc<TraitDatum> {
-        self.trait_data[&id].clone()
-    }
-
-    fn impl_datum(&self, id: ImplId) -> Arc<ImplDatum> {
-        self.impl_data[&id].clone()
-    }
-
-    fn struct_datum(&self, id: StructId) -> Arc<StructDatum> {
-        self.struct_data[&id].clone()
-    }
-
-    fn impls_for_trait(&self, trait_id: TraitId) -> Vec<ImplId> {
-        self.impl_data
-            .iter()
-            .filter(|(_, impl_datum)| {
-                let impl_trait_id = impl_datum.binders.value.trait_ref.trait_ref().trait_id;
-                impl_trait_id == trait_id
-            })
-            .map(|(&impl_id, _)| impl_id)
-            .collect()
-    }
-
-    fn impl_provided_for(&self, auto_trait_id: TraitId, struct_id: StructId) -> bool {
-        // Look for an impl like `impl Send for Foo` where `Foo` is
-        // the struct.  See `push_auto_trait_impls` for more.
-        let type_kind_id = TypeKindId::StructId(struct_id);
-        self.impl_data.values().any(|impl_datum| {
-            let impl_trait_ref = impl_datum.binders.value.trait_ref.trait_ref();
-            impl_trait_ref.trait_id == auto_trait_id
-                && match impl_trait_ref.parameters[0].assert_ty_ref() {
-                    Ty::Apply(apply) => match apply.name {
-                        TypeName::TypeKindId(id) => id == type_kind_id,
-                        _ => false,
-                    },
-
-                    _ => false,
-                }
-        })
-    }
-
-    fn type_name(&self, id: TypeKindId) -> Identifier {
-        match self.type_kinds.get(&id) {
-            Some(v) => v.name,
-            None => panic!("no type with id `{:?}`", id),
-        }
-    }
-
-    fn split_projection<'p>(
-        &self,
-        projection: &'p ProjectionTy,
-    ) -> (Arc<AssociatedTyDatum>, &'p [Parameter], &'p [Parameter]) {
-        self.split_projection(projection)
-    }
-}
-
-impl Program {
-    pub fn environment(&self) -> ProgramEnvironment {
-        // Construct the set of *clauses*; these are sort of a compiled form
-        // of the data above that always has the form:
-        //
-        //       forall P0...Pn. Something :- Conditions
-        let mut program_clauses = self.custom_clauses.clone();
-
-        self.associated_ty_data
-            .values()
-            .for_each(|d| d.to_program_clauses(self, &mut program_clauses));
-
-        self.trait_data
-            .values()
-            .for_each(|d| d.to_program_clauses(self, &mut program_clauses));
-
-        self.struct_data
-            .values()
-            .for_each(|d| d.to_program_clauses(self, &mut program_clauses));
-
-        for (&auto_trait_id, auto_trait) in self
-            .trait_data
-            .iter()
-            .filter(|(_, auto_trait)| auto_trait.binders.value.flags.auto)
-        {
-            for (&struct_id, struct_datum) in self.struct_data.iter() {
-                chalk_rules::clauses::push_auto_trait_impls(
-                    auto_trait_id,
-                    auto_trait,
-                    struct_id,
-                    struct_datum,
-                    self,
-                    &mut program_clauses,
-                );
-            }
-        }
-
-        for datum in self.impl_data.values() {
-            // If we encounter a negative impl, do not generate any rule. Negative impls
-            // are currently just there to deactivate default impls for auto traits.
-            if datum.binders.value.trait_ref.is_positive() {
-                datum.to_program_clauses(self, &mut program_clauses);
-                datum
-                    .binders
-                    .value
-                    .associated_ty_values
-                    .iter()
-                    .for_each(|atv| atv.to_program_clauses(self, &mut program_clauses));
-            }
-        }
-
-        let coinductive_traits = self
-            .trait_data
-            .iter()
-            .filter_map(|(&trait_id, trait_datum)| {
-                if trait_datum.binders.value.flags.auto {
-                    Some(trait_id)
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        ProgramEnvironment::new(coinductive_traits, program_clauses)
     }
 }
