+++ src/query.rs
@@ -7,16 +7,18 @@ use crate::program::Program;
 use crate::program_environment::ProgramEnvironment;
 use chalk_ir::tls;
 use chalk_ir::TraitId;
+use chalk_rules::clauses::ToProgramClauses;
 use chalk_rules::coherence::orphan;
 use chalk_rules::coherence::{CoherenceSolver, SpecializationPriorities};
 use chalk_rules::wf;
+use chalk_rules::RustIrSource;
 use chalk_solve::ProgramClauseSet;
 use chalk_solve::SolverChoice;
 use std::collections::BTreeMap;
 use std::sync::Arc;
 
 #[salsa::query_group(Lowering)]
-pub trait LoweringDatabase: ProgramClauseSet {
+pub trait LoweringDatabase: ProgramClauseSet + RustIrSource {
     #[salsa::input]
     fn program_text(&self) -> Arc<String>;
 
@@ -50,7 +52,7 @@ fn orphan_check(db: &impl LoweringDatabase) -> Result<(), ChalkError> {
     tls::set_current_program(&program, || -> Result<(), ChalkError> {
         let local_impls = program.local_impl_ids();
         for impl_id in local_impls {
-            orphan::perform_orphan_check(&*program, db, solver_choice, impl_id)?;
+            orphan::perform_orphan_check(db, db, solver_choice, impl_id)?;
         }
         Ok(())
     })
@@ -65,7 +67,7 @@ fn coherence(
         .trait_data
         .keys()
         .map(|&trait_id| {
-            let solver = CoherenceSolver::new(&*program, db, db.solver_choice(), trait_id);
+            let solver = CoherenceSolver::new(db, db, db.solver_choice(), trait_id);
             let priorities = solver.specialization_priorities()?;
             Ok((trait_id, priorities))
         })
@@ -84,7 +86,7 @@ fn checked_program(db: &impl LoweringDatabase) -> Result<Arc<Program>, ChalkErro
 
     let () = tls::set_current_program(&program, || -> Result<(), ChalkError> {
         let solver = wf::WfSolver {
-            program: &*program,
+            program: db,
             env: db,
             solver_choice: db.solver_choice(),
         };
@@ -104,6 +106,74 @@ fn checked_program(db: &impl LoweringDatabase) -> Result<Arc<Program>, ChalkErro
 }
 
 fn environment(db: &impl LoweringDatabase) -> Result<Arc<ProgramEnvironment>, ChalkError> {
-    let env = db.program_ir()?.environment();
-    Ok(Arc::new(env))
+    let program = db.program_ir()?;
+
+    // Construct the set of *clauses*; these are sort of a compiled form
+    // of the data above that always has the form:
+    //
+    //       forall P0...Pn. Something :- Conditions
+    let mut program_clauses = program.custom_clauses.clone();
+
+    program
+        .associated_ty_data
+        .values()
+        .for_each(|d| d.to_program_clauses(db, &mut program_clauses));
+
+    program
+        .trait_data
+        .values()
+        .for_each(|d| d.to_program_clauses(db, &mut program_clauses));
+
+    program
+        .struct_data
+        .values()
+        .for_each(|d| d.to_program_clauses(db, &mut program_clauses));
+
+    for (&auto_trait_id, auto_trait) in program
+        .trait_data
+        .iter()
+        .filter(|(_, auto_trait)| auto_trait.binders.value.flags.auto)
+    {
+        for (&struct_id, struct_datum) in program.struct_data.iter() {
+            chalk_rules::clauses::push_auto_trait_impls(
+                auto_trait_id,
+                auto_trait,
+                struct_id,
+                struct_datum,
+                db,
+                &mut program_clauses,
+            );
+        }
+    }
+
+    for datum in program.impl_data.values() {
+        // If we encounter a negative impl, do not generate any rule. Negative impls
+        // are currently just there to deactivate default impls for auto traits.
+        if datum.binders.value.trait_ref.is_positive() {
+            datum.to_program_clauses(db, &mut program_clauses);
+            datum
+                .binders
+                .value
+                .associated_ty_values
+                .iter()
+                .for_each(|atv| atv.to_program_clauses(db, &mut program_clauses));
+        }
+    }
+
+    let coinductive_traits = program
+        .trait_data
+        .iter()
+        .filter_map(|(&trait_id, trait_datum)| {
+            if trait_datum.binders.value.flags.auto {
+                Some(trait_id)
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    Ok(Arc::new(ProgramEnvironment::new(
+        coinductive_traits,
+        program_clauses,
+    )))
 }
